functions/requirements-investor-agent.txt
# Investor Recommendation Agent Dependencies
# Core Google Cloud and Firebase
firebase-functions>=0.2.0
firebase-admin>=6.5.0
google-cloud-aiplatform>=1.49.0
google-cloud-firestore>=2.13.0
google-cloud-functions-framework>=3.4.0

# AI/ML Dependencies for Investor Matching
vertexai>=1.38.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# Optional: For enhanced text processing
nltk>=3.8.1
spacy>=3.7.0
